refactor(eval): page-level data layer reads column config from store, drops legacy migration
Move DataTableStateProvider up to wrap the eval page (and shared page) so EvaluationContent reads customColumns directly via useDataTableStore selector — no more callbacks-up, no enableRealtime prop, no legacy localStorage migration. The page builds columnDefs first (config layer), then constructs URLs and runs stats SWR + useInfiniteScroll + useRealtime against them (data layer). Maps cleanly to a future where column config comes from the DB instead of localStorage: the same shape works once persistence moves.

Drop scoreRanges and isComparison from EvalStore — both are now derived in-render and threaded through tableMeta. Removes two effects (URL→isComparison sync, data→scoreRanges sync) and one store field. EvaluationDatapointsTable becomes a small presentational component with no data fetching of its own; useRealtime lives only in the non-shared page.
