+++ frontend/components/evaluation/evaluation.tsx
@@ -1,19 +1,38 @@
 "use client";
 
 import { type Row } from "@tanstack/react-table";
+import { debounce } from "lodash";
 import { useParams, usePathname, useRouter, useSearchParams } from "next/navigation";
-import { useCallback, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import useSWR from "swr";
+import { useStore } from "zustand";
+import { shallow } from "zustand/shallow";
 
 import Chart from "@/components/evaluation/chart";
 import CompareChart from "@/components/evaluation/compare-chart";
 import EvaluationDatapointsTable from "@/components/evaluation/evaluation-datapoints-table";
 import EvaluationHeader from "@/components/evaluation/evaluation-header";
 import ScoreCard from "@/components/evaluation/score-card";
-import { EvalStoreProvider } from "@/components/evaluation/store";
-import { type EvaluationStatsPayload } from "@/components/evaluation/utils";
+import {
+  buildColumnDefs,
+  buildFetchParams,
+  buildStatsParams,
+  EvalStoreProvider,
+  selectVisibleColumnDefs,
+  useEvalStore,
+} from "@/components/evaluation/store";
+import {
+  type EvaluationStatsPayload,
+  flattenScores,
+  mergeDatapointUpsertIntoRows,
+  mergeTraceUpdateIntoRows,
+} from "@/components/evaluation/utils";
+import { useInfiniteScroll } from "@/components/ui/infinite-datatable/hooks";
+import { DataTableStateProvider, useDataTableStore } from "@/components/ui/infinite-datatable/model/datatable-store";
 import { Skeleton } from "@/components/ui/skeleton";
-import { type EvalRow, type Evaluation as EvaluationType } from "@/lib/evaluation/types";
-import { formatTimestamp } from "@/lib/utils";
+import { type EvalRow, type Evaluation as EvaluationType, type EvaluationResultsInfo } from "@/lib/evaluation/types";
+import { useRealtime } from "@/lib/hooks/use-realtime";
+import { formatTimestamp, swrFetcher } from "@/lib/utils";
 
 import { TraceViewSidePanel } from "../traces/trace-view";
 import Header from "../ui/header";
@@ -26,43 +45,179 @@ interface EvaluationProps {
   initialScoreNames: string[];
 }
 
-function EvaluationContent({ evaluations, evaluationId, evaluationName, initialScoreNames }: EvaluationProps) {
+const PAGE_SIZE = 50;
+const BASE_COLUMN_ORDER = ["status", "index", "data", "target", "metadata", "output", "duration", "cost"];
+
+function EvaluationContent({ evaluations, evaluationId, evaluationName }: EvaluationProps) {
   const { push } = useRouter();
   const pathName = usePathname();
   const searchParams = useSearchParams();
   const params = useParams<{ projectId: string }>();
+
+  const search = searchParams.get("search");
+  const filter = searchParams.getAll("filter");
+  const sortBy = searchParams.get("sortBy");
+  const sortDirection = searchParams.get("sortDirection");
   const targetId = searchParams.get("targetId");
 
-  const [selectedScore, setSelectedScore] = useState<string | undefined>(() => initialScoreNames[0]);
+  // Column config layer: customColumns are read from the DataTableStore (the
+  // persisted-config layer) and threaded into the columnDefs / URLs below.
+  const datatableStore = useDataTableStore<EvalRow>();
+  const { customColumns, removeCustomColumn } = useStore(
+    datatableStore,
+    (s) => ({ customColumns: s.customColumns, removeCustomColumn: s.removeCustomColumn }),
+    shallow
+  );
+
+  // Eval-specific state lives in EvalStore. customColumns intentionally do not.
+  const scoreNames = useEvalStore((s) => s.scoreNames);
+  const isShared = useEvalStore((s) => s.isShared);
+  const heatmapEnabled = useEvalStore((s) => s.heatmapEnabled);
+  const setHeatmapEnabled = useEvalStore((s) => s.setHeatmapEnabled);
+  const addScoreName = useEvalStore((s) => s.addScoreName);
+
+  const isComparison = !!targetId;
+  const columnDefs = useMemo(
+    () => buildColumnDefs({ scoreNames, customColumns, isShared }),
+    [scoreNames, customColumns, isShared]
+  );
+
+  // Stats SWR — drives the score card + chart.
+  const statsUrl = useMemo(() => {
+    const base = `/api/projects/${params.projectId}/evaluations/${evaluationId}/stats`;
+    const urlParams = buildStatsParams({ search, filter, sortBy, sortDirection }, columnDefs, scoreNames);
+    const qs = urlParams.toString();
+    return qs ? `${base}?${qs}` : base;
+  }, [params.projectId, evaluationId, search, filter, sortBy, sortDirection, columnDefs, scoreNames]);
+
+  const {
+    data: statsData,
+    isLoading: isStatsLoading,
+    mutate: mutateStats,
+  } = useSWR<EvaluationStatsPayload>(statsUrl, swrFetcher, { revalidateOnFocus: false });
+
+  const targetStatsUrl = useMemo(() => {
+    if (!targetId) return null;
+    const base = `/api/projects/${params.projectId}/evaluations/${targetId}/stats`;
+    const urlParams = buildStatsParams({ search, filter, sortBy, sortDirection }, columnDefs, scoreNames);
+    const qs = urlParams.toString();
+    return qs ? `${base}?${qs}` : base;
+  }, [params.projectId, targetId, search, filter, sortBy, sortDirection, columnDefs, scoreNames]);
+
+  const { data: targetStatsData } = useSWR<EvaluationStatsPayload>(targetStatsUrl, swrFetcher, {
+    revalidateOnFocus: false,
+  });
+
+  // Datapoints fetcher — depends on columnDefs (custom column SQL, etc).
+  // SQL strings are stable across cosmetic columnDefs changes; JSON.stringify
+  // on `columnSqls` produces the same string → no spurious refetch.
+  const columnSqls = useMemo(() => columnDefs.map((c) => c.meta?.sql).filter(Boolean), [columnDefs]);
+
+  const fetchDatapoints = useCallback(
+    async (pageNumber: number) => {
+      const urlParams = buildFetchParams(
+        { search, filter, sortBy, sortDirection, targetId, pageNumber, pageSize: PAGE_SIZE },
+        columnDefs
+      );
+      const url = `/api/projects/${params.projectId}/evaluations/${evaluationId}?${urlParams.toString()}`;
+      const response = await fetch(url);
+      if (!response.ok) throw new Error("Failed to fetch datapoints.");
+      const data: EvaluationResultsInfo = await response.json();
+      return { items: data.results, count: 0 };
+    },
+    [search, filter, params.projectId, evaluationId, sortBy, sortDirection, targetId, columnDefs]
+  );
+
+  const {
+    data: allDatapoints,
+    hasMore,
+    isFetching,
+    isLoading: isLoadingDatapoints,
+    fetchNextPage,
+    updateData,
+  } = useInfiniteScroll<EvalRow>({
+    fetchFn: fetchDatapoints,
+    enabled: !isStatsLoading,
+    deps: [search, filter, evaluationId, sortBy, sortDirection, targetId, columnSqls],
+  });
+
+  // Score-range heatmap input — derived from current data, no storage needed.
+  const scoreRanges = useMemo(() => {
+    if (!allDatapoints) return {};
+    const isValidNumber = (value: unknown): value is number => typeof value === "number" && !isNaN(value);
+    return scoreNames.reduce(
+      (acc, scoreName) => {
+        const values = allDatapoints
+          .flatMap((row) => {
+            const v = [row[`score:${scoreName}`]];
+            if (targetId) v.push(row[`compared:score:${scoreName}`]);
+            return v;
+          })
+          .filter(isValidNumber);
+        if (values.length === 0) return acc;
+        return { ...acc, [scoreName]: { min: Math.min(...values), max: Math.max(...values) } };
+      },
+      {} as Record<string, { min: number; max: number }>
+    );
+  }, [allDatapoints, scoreNames, targetId]);
+
+  // Realtime — only on the live (non-comparison) eval page.
+  const debouncedRevalidateStats = useMemo(
+    () => debounce(() => mutateStats(), 1000, { leading: false, trailing: true }),
+    [mutateStats]
+  );
+  useEffect(() => () => debouncedRevalidateStats.cancel(), [debouncedRevalidateStats]);
+
+  const realtimeHandlers = useMemo(
+    () => ({
+      datapoint_upsert: (event: MessageEvent) => {
+        try {
+          const payload = JSON.parse(event.data) as { datapoints?: Array<EvalRow & { id: string }> };
+          payload.datapoints?.forEach((incoming) => {
+            const flattened = flattenScores(incoming["scores"]);
+            updateData((rows) => mergeDatapointUpsertIntoRows(rows, incoming, flattened));
+            if (Object.keys(flattened).length === 0) return;
+            Object.keys(flattened).forEach((key) => addScoreName(key.slice("score:".length)));
+            debouncedRevalidateStats();
+          });
+        } catch (e) {
+          console.warn("Failed to parse realtime datapoint_upsert:", e);
+        }
+      },
+      trace_update: (event: MessageEvent) => {
+        try {
+          const payload = JSON.parse(event.data) as { traces?: Array<Record<string, unknown> & { id: string }> };
+          payload.traces?.forEach((trace) => updateData((rows) => mergeTraceUpdateIntoRows(rows, trace)));
+        } catch (e) {
+          console.warn("Failed to parse realtime trace_update:", e);
+        }
+      },
+    }),
+    [updateData, addScoreName, debouncedRevalidateStats]
+  );
+
+  useRealtime({
+    key: `evaluation_${evaluationId}`,
+    projectId: params.projectId,
+    enabled: !targetId,
+    eventHandlers: realtimeHandlers,
+  });
+
+  // Side-panel + selected-row state for trace view.
+  const [selectedScore, setSelectedScore] = useState<string | undefined>(() => scoreNames[0]);
   const [traceId, setTraceId] = useState<string | undefined>(() => searchParams.get("traceId") ?? undefined);
   const [datapointId, setDatapointId] = useState<string | undefined>(
     () => searchParams.get("datapointId") ?? undefined
   );
-  const [selectedRow, setSelectedRow] = useState<EvalRow | undefined>(undefined);
-  const [statsData, setStatsData] = useState<EvaluationStatsPayload | undefined>(undefined);
-  const [targetStatsData, setTargetStatsData] = useState<EvaluationStatsPayload | undefined>(undefined);
 
-  const isStatsLoading = statsData === undefined;
+  if (!selectedScore && scoreNames.length > 0) {
+    setSelectedScore(scoreNames[0]);
+  }
 
-  const buildDatapointsUrl = useCallback(
-    (qs: string) => `/api/projects/${params.projectId}/evaluations/${evaluationId}?${qs}`,
-    [params.projectId, evaluationId]
+  const selectedRow = useMemo<EvalRow | undefined>(
+    () => allDatapoints?.find((row) => row["id"] === datapointId),
+    [allDatapoints, datapointId]
   );
-  const buildStatsUrl = useCallback(
-    (qs: string) => {
-      const base = `/api/projects/${params.projectId}/evaluations/${evaluationId}/stats`;
-      return qs ? `${base}?${qs}` : base;
-    },
-    [params.projectId, evaluationId]
-  );
-
-  const onClose = useCallback(() => {
-    setTraceId(undefined);
-    const next = new URLSearchParams(searchParams.toString());
-    next.delete("traceId");
-    next.delete("spanId");
-    push(`${pathName}?${next}`);
-  }, [searchParams, pathName, push]);
 
   const handleRowClick = useCallback((row: Row<EvalRow>) => {
     setTraceId(row.original["traceId"] as string);
@@ -79,6 +234,29 @@ function EvaluationContent({ evaluations, evaluationId, evaluationName, initialS
     [pathName, searchParams]
   );
 
+  const handleSort = useCallback(
+    (columnId: string, direction: "asc" | "desc") => {
+      const next = new URLSearchParams(searchParams.toString());
+      if (columnId) {
+        next.set("sortBy", columnId);
+        next.set("sortDirection", direction.toUpperCase());
+      } else {
+        next.delete("sortBy");
+        next.delete("sortDirection");
+      }
+      push(`${pathName}?${next.toString()}`);
+    },
+    [searchParams, push, pathName]
+  );
+
+  const onClose = useCallback(() => {
+    setTraceId(undefined);
+    const next = new URLSearchParams(searchParams.toString());
+    next.delete("traceId");
+    next.delete("spanId");
+    push(`${pathName}?${next}`);
+  }, [searchParams, pathName, push]);
+
   const handleTraceChange = (id: string) => {
     const next = new URLSearchParams(searchParams);
     next.set("traceId", id);
@@ -86,17 +264,15 @@ function EvaluationContent({ evaluations, evaluationId, evaluationName, initialS
     setTraceId(id);
   };
 
-  const scoreNames = statsData?.evaluation
-    ? Object.keys(statsData.allStatistics ?? {}).length > 0
-      ? Object.keys(statsData.allStatistics)
-      : initialScoreNames
-    : initialScoreNames;
+  const visibleColumnDefs = useMemo(
+    () => selectVisibleColumnDefs(columnDefs, isComparison),
+    [columnDefs, isComparison]
+  );
 
-  if (!selectedScore && scoreNames.length > 0) {
-    setSelectedScore(scoreNames[0]);
-  }
-
-  const statsUrl = useMemo(() => buildStatsUrl(""), [buildStatsUrl]);
+  const onDeleteCustomColumn = useCallback(
+    (columnId: string) => removeCustomColumn(columnId.replace("custom:", "")),
+    [removeCustomColumn]
+  );
 
   return (
     <>
@@ -150,16 +326,24 @@ function EvaluationContent({ evaluations, evaluationId, evaluationName, initialS
             )}
           </div>
           <EvaluationDatapointsTable
-            evaluationId={evaluationId}
-            initialScoreNames={initialScoreNames}
-            buildDatapointsUrl={buildDatapointsUrl}
-            buildStatsUrl={buildStatsUrl}
+            data={allDatapoints}
+            isLoading={isStatsLoading || isLoadingDatapoints}
+            isFetching={isFetching}
+            hasMore={hasMore}
+            fetchNextPage={fetchNextPage}
+            columnDefs={columnDefs}
+            visibleColumnDefs={visibleColumnDefs}
+            isComparison={isComparison}
+            scoreRanges={scoreRanges}
+            datapointId={datapointId}
             handleRowClick={handleRowClick}
             getRowHref={getRowHref}
-            datapointId={datapointId}
-            onStatsLoaded={setStatsData}
-            onTargetStatsLoaded={setTargetStatsData}
-            onSelectedRowChange={setSelectedRow}
+            sortBy={sortBy ?? undefined}
+            sortDirection={(sortDirection?.toLowerCase() ?? undefined) as "asc" | "desc" | undefined}
+            onSort={handleSort}
+            heatmapEnabled={heatmapEnabled}
+            onHeatmapEnabledChange={setHeatmapEnabled}
+            onDeleteCustomColumn={onDeleteCustomColumn}
           />
         </div>
       </div>
@@ -203,9 +387,16 @@ function EvaluationContent({ evaluations, evaluationId, evaluationName, initialS
 }
 
 export default function Evaluation(props: EvaluationProps) {
+  const defaultColumnOrder = useMemo(
+    () => [...BASE_COLUMN_ORDER, ...props.initialScoreNames.map((s) => `score:${s}`)],
+    [props.initialScoreNames]
+  );
+
   return (
     <EvalStoreProvider key={props.evaluationId} initialScoreNames={props.initialScoreNames}>
-      <EvaluationContent {...props} />
+      <DataTableStateProvider storageKey="evaluation-datapoints-table" defaultColumnOrder={defaultColumnOrder}>
+        <EvaluationContent {...props} />
+      </DataTableStateProvider>
     </EvalStoreProvider>
   );
 }
